Separate recipe icons component
