+++ app/ui/content/recipe-card.tsx
@@ -1,13 +1,10 @@
 "use client";
 
-import { EllipsisVerticalIcon  } from "@heroicons/react/24/solid";
-import { HeartIcon as EmptySaveIcon } from "@heroicons/react/24/outline";
-import { HeartIcon as FilledSaveIcon } from "@heroicons/react/24/solid";
 import { LargeRecipeImage, SmallRecipeImage } from "@/app/ui/content/recipe-image";
 import React from 'react';
 import {RecipeOverview} from "@/app/lib/recipeOverview";
 import Link from "next/link";
-import { isRecipeSaved, saveRecipe, removeRecipe } from "@/app/lib/savedRecipes";
+import {RecipeIcons} from "@/app/ui/navigation/recipe-icons";
 
 export type RecipeCardData = {
     recipe: RecipeOverview,
@@ -41,7 +38,7 @@ function Body({data, onOptionsClick}: { data: RecipeCardData, onOptionsClick?: (
             <Link href={"/recipes/" + data.recipe.recipe_id}>
                 <RecipeInfo data={data} />
             </Link>
-            <Icons data={data} onOptionsClick={onOptionsClick} />
+            <RecipeIcons recipe_id={data.recipe.recipe_id} styling={"p-2 pt-6"} onOptionsClick={onOptionsClick} />
         </div>
     )
 }
@@ -62,41 +59,4 @@ function RecipeText({data}: { data: RecipeCardData }) {
           <p>{data.recipe.user_name}</p>
         </div>
     )
-}
-
-function Icons({data, onOptionsClick}: {data: RecipeCardData, onOptionsClick?: () => void}) {
-    const isSaved = isRecipeSaved(data.recipe.recipe_id);
-    const [saveButtonEnabled, setSaveButtonEnabled] = React.useState(isSaved);
-    const [isClient, setIsClient] = React.useState(false)
-    React.useEffect(() => {
-        setIsClient(true)
-    }, [])
-
-    return (
-        <div className="flex flex-row items-center p-2 pt-6">
-            <button
-                onClick={() => onSavedRecipePress(data.recipe.recipe_id, setSaveButtonEnabled)}>
-                {isClient && saveButtonEnabled ?
-                    <FilledSaveIcon className="size-8"/> :
-                    <EmptySaveIcon className="size-8"/>
-                }
-            </button>
-            { onOptionsClick != null ?
-                <button onClick={onOptionsClick}>
-                    <EllipsisVerticalIcon className="size-8"/>
-                </button> :
-                null
-            }
-        </div>
-    )
-}
-
-function onSavedRecipePress(recipe_id: number, setSaveButtonState: React.Dispatch<React.SetStateAction<boolean>>) {
-    if (isRecipeSaved(recipe_id)) {
-        removeRecipe(recipe_id)
-        setSaveButtonState(false)
-    } else {
-        saveRecipe(recipe_id)
-        setSaveButtonState(true)
-    }
 }
